Removing Lifetime parameter
Removing lifetime parameter simplify the usage of the session as there
will not be the need to define a lifetime for the session as well as
making the code more readable.

This can be done by cloning the environment which is an Arc Mutex so not
costly to clone.

This PR enables the use of onnxruntime-rs session within actix-web
threads.